Create form-utils to abstract shared form logic
